+++ src/Factories/RequestFactory.php
@@ -330,6 +330,60 @@ abstract class RequestFactory
     /**
      * @throws EbicsException
      */
+    public function createHAC(
+        DateTimeInterface $dateTime,
+        int $segmentNumber = null,
+        bool $isLastSegment = null
+    ): Request {
+        $context = (new RequestContext())
+            ->setBank($this->bank)
+            ->setUser($this->user)
+            ->setKeyRing($this->keyRing)
+            ->setDateTime($dateTime)
+            ->setSegmentNumber($segmentNumber)
+            ->setIsLastSegment($isLastSegment);
+
+        $request = $this
+            ->createRequestBuilderInstance()
+            ->addContainerSecured(function (XmlBuilder $builder) use ($context) {
+                $builder->addHeader(function (HeaderBuilder $builder) use ($context) {
+                    $builder->addStatic(function (StaticBuilder $builder) use ($context) {
+                        $builder
+                            ->addHostId($context->getBank()->getHostId())
+                            ->addRandomNonce()
+                            ->addTimestamp($context->getDateTime())
+                            ->addPartnerId($context->getUser()->getPartnerId())
+                            ->addUserId($context->getUser()->getUserId())
+                            ->addProduct('Ebics client PHP', 'de')
+                            ->addOrderDetails(function (OrderDetailsBuilder $orderDetailsBuilder) {
+                                $this
+                                    ->addOrderType($orderDetailsBuilder, 'HAC')
+                                    ->addStandardOrderParams();
+                            })
+                            ->addBankPubKeyDigests(
+                                $context->getKeyRing()->getBankSignatureXVersion(),
+                                $this->digestResolver->digest($context->getKeyRing()->getBankSignatureX()),
+                                $context->getKeyRing()->getBankSignatureEVersion(),
+                                $this->digestResolver->digest($context->getKeyRing()->getBankSignatureE())
+                            )
+                            ->addSecurityMedium(StaticBuilder::SECURITY_MEDIUM_0000);
+                    })->addMutable(function (MutableBuilder $builder) use ($context) {
+                        $builder
+                            ->addTransactionPhase(MutableBuilder::PHASE_INITIALIZATION)
+                            ->addSegmentNumber($context->getSegmentNumber(), $context->getIsLastSegment());
+                    });
+                })->addBody();
+            })
+            ->popInstance();
+
+        $this->authSignatureHandler->handle($request);
+
+        return $request;
+    }
+
+    /**
+     * @throws EbicsException
+     */
     public function createHKD(
         DateTimeInterface $dateTime,
         int $segmentNumber = null,
@@ -852,6 +906,8 @@ abstract class RequestFactory
     abstract public function createCCT(DateTimeInterface $dateTime, UploadTransaction $transaction): Request;
 
     abstract public function createCDD(DateTimeInterface $dateTime, UploadTransaction $transaction): Request;
+
+    abstract public function createCDS(DateTimeInterface $dateTime, UploadTransaction $transaction): Request;
 
     abstract public function createXE2(DateTimeInterface $dateTime, UploadTransaction $transaction): Request;